# Msgpack perf: write psMap out directly