--- a/report/edge_metadatas_internal_test.go
+++ b/report/edge_metadatas_internal_test.go
@@ -215,18 +215,6 @@ func TestEdgeMetadatasEncoding(t *testing.T) {
 		})
 
 	{
-		gobs, err := want.GobEncode()
-		if err != nil {
-			t.Fatal(err)
-		}
-		have := EmptyEdgeMetadatas
-		have.GobDecode(gobs)
-		if !reflect.DeepEqual(want, have) {
-			t.Error(test.Diff(want, have))
-		}
-	}
-
-	{
 		for _, h := range []codec.Handle{
 			codec.Handle(&codec.MsgpackHandle{}),
 			codec.Handle(&codec.JsonHandle{}),
@@ -246,18 +234,6 @@ func TestEdgeMetadatasEncoding(t *testing.T) {
 
 func TestEdgeMetadatasEncodingNil(t *testing.T) {
 	want := EdgeMetadatas{}
-
-	{
-		gobs, err := want.GobEncode()
-		if err != nil {
-			t.Fatal(err)
-		}
-		have := EmptyEdgeMetadatas
-		have.GobDecode(gobs)
-		if have.psMap == nil {
-			t.Error("needed to get back a non-nil psMap for EdgeMetadata")
-		}
-	}
 
 	{