There is a big performance improvement from sending the data directly rather than creating an intermediate data structure.

Before:
```
vagrant-ubuntu-wily-64:~/.../weaveworks/scope/report$ go test -cpu=1 -run=EncodeSpeed
ok  	github.com/weaveworks/scope/report	3.978s
ok  	github.com/weaveworks/scope/report	4.035s
ok  	github.com/weaveworks/scope/report	3.945s
ok  	github.com/weaveworks/scope/report	4.184s
```
After:
```
ok  	github.com/weaveworks/scope/report	1.951s
ok  	github.com/weaveworks/scope/report	1.934s
ok  	github.com/weaveworks/scope/report	1.971s
ok  	github.com/weaveworks/scope/report	1.970s
```

1.934/3.945=49%

Test program:

```
func TestEncodeSpeed(t *testing.T) {
	b, _ := ioutil.ReadFile("fixed.msgpack")
	r := Report{}
	d := codec.NewDecoderBytes(b, &codec.MsgpackHandle{})
	d.Decode(&r)
	for i := 0; i < 100; i++ {
		buf := &bytes.Buffer{}
		codec.NewEncoder(buf, &codec.MsgpackHandle{}).Encode(&r)
	}
}
```