@bboreham 2017-04-25
Note that some `toIntermediate` functions remain, because they are called by Gob code.  Is Gob used anywhere in Scope nowadays?