@2opremio 2017-04-25
> Note that some toIntermediate functions remain, because they are called by Gob code. Is Gob used anywhere in Scope nowadays?

No, please ditch them (https://github.com/weaveworks/scope/issues/971)