@bboreham 2017-04-25
gob ditched